rename indicator dir
