+++ src/ui/components/badge/ImageIndicator.tsx
