+++ src/ui/components/badge/TopRightIndicator.tsx
@@ -1,10 +1,10 @@
 import { PKMInterface } from '@openhome-core/pkm/interfaces'
 import { OHPKM } from '@openhome-core/pkm/OHPKM'
+import { TopRightIndicatorType } from '@openhome-ui/hooks/monDisplay'
+import { getPublicImageURL } from '@openhome-ui/images/images'
+import { BallsImageList } from '@openhome-ui/images/items'
 import { Pokerus, StatsPreSplit } from '@pkm-rs/pkg/pkm_rs'
-import { TopRightIndicatorType } from '../../../hooks/monDisplay'
-import { getPublicImageURL } from '../../../images/images'
-import { BallsImageList } from '../../../images/items'
-import GenderIcon from '../GenderIcon'
+import GenderIcon from '../pokemon/GenderIcon'
 import { GameIndicator } from './GameIndicator'
 import { ImageIndicator } from './ImageIndicator'
 import { Indicator } from './Indicator'
