+++ src/ui/pages/tracked/TrackedPokemonPage.tsx
@@ -2,9 +2,9 @@ import { PKMInterface } from '@openhome-core/pkm/interfaces'
 import { OhpkmIdentifier } from '@openhome-core/pkm/Lookup'
 import { SAV } from '@openhome-core/save/interfaces'
 import { numericSorter } from '@openhome-core/util/sort'
+import { GameIndicator } from '@openhome-ui/components/badge/GameIndicator'
 import { Dialog } from '@openhome-ui/components/dialog/Dialog'
 import MessageRibbon from '@openhome-ui/components/MessageRibbon'
-import { GameIndicator } from '@openhome-ui/components/pokemon/indicator/GameIndicator'
 import SideTabNavigation from '@openhome-ui/components/side-tabs/SideTabNavigation'
 import PokemonDetailsModal from '@openhome-ui/pokemon-details/Modal'
 import { useSaves } from '@openhome-ui/state/saves'
