+++ src/ui/pokemon-details/tabs/SummaryTab.tsx
@@ -6,9 +6,9 @@ import { getDisplayID } from '@openhome-core/util'
 import { pluralize } from '@openhome-core/util/format'
 import { Option } from '@openhome-core/util/functional'
 import AttributeRow from '@openhome-ui/components/AttributeRow'
+import { ImageIndicator } from '@openhome-ui/components/badge/ImageIndicator'
 import { ErrorIcon } from '@openhome-ui/components/Icons'
 import GenderIcon from '@openhome-ui/components/pokemon/GenderIcon'
-import { ImageIndicator } from '@openhome-ui/components/pokemon/indicator/ImageIndicator'
 import TypeIcon from '@openhome-ui/components/pokemon/TypeIcon'
 import PokemonIcon from '@openhome-ui/components/PokemonIcon'
 import { Tabs } from '@openhome-ui/components/Tabs'
